Rename `loader.js` to `Loadable.js`
